requirements.txt
camelot_py==0.11.0
pandas==2.0.3
